Move files, add to documentation
