--- ott/core/gw_barycenter/__init__.py
@@ -1,2 +1,0 @@
-from .problem import GWBarycenterProblem
-from .solver import GromovWassersteinBarycenter
